refactor(web): mv lists/selectIcon to components/selectIcons
